refactor: turns `translations-history` into a folder and uses `TranslationsHistoryContent` component
